+++ packages/backend/src/peripherals/starkware/schema.ts
@@ -4,7 +4,7 @@ const UnsignedIntAsString = z.string().regex(/^([1-9]\d*|0)$/)
 const SignedIntAsString = z.string().regex(/^(-?[1-9]\d*|0)$/)
 const PedersenHash = z.string().regex(/^0[a-f\d]{63}$/)
 const PedersenHash0x = z.string().regex(/^0x[a-f\d]{0,64}$/)
-const Hash256 = z.string().regex(/^0x[a-f\d]{64}$/)
+const Hash256 = z.string().regex(/^0x[a-f\d]{0,64}$/)
 const StarkKey = z.string().regex(/^0x0[a-f\d]{63}$/)
 const Bytes32 = z.string().regex(/^[a-f\d]{64}$/)
 const AssetHash0x = z.string().regex(/^0x[a-f\d]{0,63}$/)
