+++ packages/backend/src/peripherals/starkware/toSpotBatchData.ts
